algoritme voor verdeling werkend

+++ back_end/src/main/java/nl/cimsolutions/snel_transport/services/AbstractFacade.java
@@ -107,12 +107,12 @@ public abstract class AbstractFacade<T> {
         return q.getResultList();
     }
 
-    public int count() {
+    public Long count() {
         javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
         javax.persistence.criteria.Root<T> rt = cq.from(entityClass);
         cq.select(getEntityManager().getCriteriaBuilder().count(rt));
         javax.persistence.Query q = getEntityManager().createQuery(cq);
-        return ((Long) q.getSingleResult()).intValue();
+        return ((Long) q.getSingleResult());
     }
 
     public void emptyTable() {
